requirements.txt
aiogram==3.17.0
aiohttp==3.11.12
pydantic==2.10.6
redis==5.2.1
python_dotenv==1.0.1
pydantic-settings==2.7.1
pre-commit==4.1.0